Refactoring of module layout
