+++ examples/rig_arm.ma
