+++ examples/rig_example_01.ma
