+++ examples/rig_example_02.ma
