+++ examples/rig_example_02_skel.ma
