+++ examples/rig_example_02_skel_new.ma
