+++ examples/rig_splineik.ma
