+++ omtk/classNode.py
