+++ omtk/classRig.py
