--- omtk/rigging/__init__.py
@@ -1,3 +1,0 @@
-import os, sys, re
-if re.match('maya.*', os.path.basename(sys.executable), re.IGNORECASE):
-    import autorig
