--- omtk/rigging/autorig/ui.ui
@@ -1,101 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ui version="4.0">
- <class>MainWindow</class>
- <widget class="QMainWindow" name="MainWindow">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>597</width>
-    <height>417</height>
-   </rect>
-  </property>
-  <property name="windowTitle">
-   <string>MainWindow</string>
-  </property>
-  <widget class="QWidget" name="centralwidget">
-   <layout class="QVBoxLayout" name="verticalLayout">
-    <item>
-     <widget class="QTreeWidget" name="treeWidget">
-      <column>
-       <property name="text">
-        <string notr="true">1</string>
-       </property>
-      </column>
-     </widget>
-    </item>
-   </layout>
-  </widget>
-  <widget class="QMenuBar" name="menubar">
-   <property name="geometry">
-    <rect>
-     <x>0</x>
-     <y>0</y>
-     <width>597</width>
-     <height>22</height>
-    </rect>
-   </property>
-  </widget>
-  <widget class="QStatusBar" name="statusbar"/>
-  <widget class="QToolBar" name="toolBar">
-   <property name="windowTitle">
-    <string>toolBar</string>
-   </property>
-   <attribute name="toolBarArea">
-    <enum>TopToolBarArea</enum>
-   </attribute>
-   <attribute name="toolBarBreak">
-    <bool>false</bool>
-   </attribute>
-   <addaction name="actionBuild"/>
-   <addaction name="actionUnbuild"/>
-   <addaction name="actionRebuild"/>
-   <addaction name="actionAdd"/>
-   <addaction name="separator"/>
-   <addaction name="actionImport"/>
-   <addaction name="actionExport"/>
-   <addaction name="separator"/>
-   <addaction name="actionUpdate"/>
-  </widget>
-  <action name="actionBuild">
-   <property name="text">
-    <string>Build</string>
-   </property>
-  </action>
-  <action name="actionUnbuild">
-   <property name="text">
-    <string>Unbuild</string>
-   </property>
-  </action>
-  <action name="actionRebuild">
-   <property name="text">
-    <string>Rebuild</string>
-   </property>
-  </action>
-  <action name="actionImport">
-   <property name="text">
-    <string>Import</string>
-   </property>
-  </action>
-  <action name="actionExport">
-   <property name="text">
-    <string>Export</string>
-   </property>
-  </action>
-  <action name="actionUpdate">
-   <property name="text">
-    <string>Update</string>
-   </property>
-   <property name="toolTip">
-    <string>Update</string>
-   </property>
-  </action>
-  <action name="actionAdd">
-   <property name="text">
-    <string>Add</string>
-   </property>
-  </action>
- </widget>
- <resources/>
- <connections/>
-</ui>
